Refactor - <App /> to host <NavMenu />
