+++ src/demo/features/index.js
@@ -1,0 +1,1 @@
+export { default as NavMenu } from './NavMenu';
