+++ src/demo/routes/components/Home/Home.test.js
@@ -14,32 +14,7 @@ describe('<Home />', () => {
     ReactDOM.render(<Home />, div);
   });
 
-  // STATE
-
-  it('should have a default mobileNavActive state', () => {
-    const state = wrapper.state().mobileNavActive;
-    expect(state).not.toBeNull();
-    expect(state).toBeDefined();
-    expect(state).not.toEqual(false);
-    expect(state).toEqual(true); // JSDOM === desktop
-  });
-
   // COMPONENT
-
-  it('should update the state to false when called', () => {
-    const func = wrapper.instance()._hideNav;
-
-    expect(func).not.toBeNull();
-    expect(func).toBeDefined();
-    expect(() => func()).not.toThrow();
-
-    wrapper.setState({ mobileNavActive: true });
-
-    func();
-    setTimeout(() => {
-      expect(wrapper.state().mobileNavActive).toEqual(false);
-    }, 400);
-  });
 
   it('should add a "fadeIn" class to img when called', () => {
     const func = wrapper.instance()._loadImage;
@@ -61,21 +36,6 @@ describe('<Home />', () => {
 
     expect(wrapper.find('.Home-whiplash.fadeIn').length).not.toEqual(0);
     expect(wrapper.find('.Home-whiplash.fadeIn').length).toEqual(1);
-  });
-
-  it('should update the state to true when called', () => {
-    const func = wrapper.instance()._showNav;
-
-    expect(func).not.toBeNull();
-    expect(func).toBeDefined();
-    expect(() => func()).not.toThrow();
-
-    wrapper.setState({ mobileNavActive: false });
-
-    func();
-    setTimeout(() => {
-      expect(wrapper.state().mobileNavActive).toEqual(true);
-    }, 400);
   });
 
 });
