+++ src/demo/routes/components/Install/Install.css
@@ -1,0 +1,3 @@
+.Install {
+  
+}
