+++ src/demo/routes/components/Install/index.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+import './Install.css';
+
+
+export default class Install extends Component {
+  render() {
+    return(
+      <div className="Install">
+        install view
+      </div>
+    );
+  }
+}
