+++ src/demo/routes/index.js
@@ -1,1 +1,2 @@
 export { default as Home } from './components/Home';
+export { default as Install } from './components/Install';
